+++ semcore/feature-highlight/src/components/tab-line/TabLine.type.ts
@@ -1,0 +1,8 @@
+import type { Intergalactic } from '@semcore/core';
+import type TabLine from '@semcore/tab-line';
+
+export type HighlightedTabLineComponent = typeof TabLine & {
+  HighlightedItem: typeof TabLine.Item & {
+    Addon: Intergalactic.Component<'div', { animatedSparkleCount?: number }>;
+  };
+};