+++ semcore/feature-highlight/src/inner-components/sparkle/AnimatedSparkles.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import Sparkle from './Sparkle';
+
+export type AnimatedSparklesProps = {
+  count?: number;
+  curve?: number;
+  top?: string;
+  left?: string;
+};
+
+export function AnimatedSparkles(props: AnimatedSparklesProps & { show: boolean }) {
+  const { count, curve, show, top, left } = props;
+
+  return show && count !== undefined && [...new Array(count)].map((_, index) => {
+    return (
+      <Sparkle key={index} index={index} num={count} curve={curve} top={top} left={left} />
+    );
+  });
+}