+++ semcore/feature-highlight/src/inner-components/sparkle/index.ts
@@ -1,0 +1,5 @@
+import Sparkle from './Sparkle';
+
+export {
+  Sparkle,
+};