+++ stories/components/data-table/tests/examples/accordion-tests/accordion-with-inline-edit.tsx
@@ -65,6 +65,7 @@ const Demo = () => {
       columns={[
         { name: 'keyword', children: 'Keyword', gtcWidth: 'minmax(60%, 80%)' },
         {
+          name: 'group1',
           children: 'Organic Sessions',
           borders: 'both',
           columns: [