+++ semcore/base-trigger/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.2.3] - 2025-07-03
+## [16.2.4] - 2025-07-07
 
 ### Changed
 
 - Inconsistent border color in disabled state for `neighborLocation=left`.
+
+## [16.2.3] - 2025-07-04
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/icon` [16.2.2 ~> 16.2.3], `@semcore/tooltip` [16.0.3 ~> 16.0.4], `@semcore/base-components` [16.1.0 ~> 16.1.1]).
 
 ## [16.2.2] - 2025-06-23